# docs/swagger.yaml
basePath: /
definitions:
  common.Response:
    properties:
      code:
        description: 自定义状态码
        type: integer
      data:
        description: 数据内容
      msg:
        description: 提示信息
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      password:
        minLength: 6
        type: string
      updatedAt:
        type: string
      username:
        description: Gin 参数校验
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: 这是一个基于 Gin 的 CRUD 示例项目
  title: Gin CRUD API
  version: "1.0"
paths:
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据 ID 删除用户
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 删除用户
      tags:
      - users
    get:
      consumes:
      - application/json
      description: 根据 ID 获取用户信息
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 获取用户详情
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 根据 ID 更新用户信息
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 更新用户
      tags:
      - users
swagger: "2.0"